+++ src/declaration/index.d.ts
@@ -4,7 +4,9 @@
  * Other properties and custom element's callbacks you can define as
  * usual.
  */
-export abstract class WebComponent<T extends State = State> extends HTMLElement {
+export abstract class WebComponent<
+  T extends object = object
+> extends HTMLElement {
   /**
    * When overriding constructor always call **super()** at start,
    * so that the correct prototype chain will be established.
@@ -146,11 +148,8 @@ export function translate(
   variables?: { [key: string]: string }
 ): string;
 
-/** Common shape of state for application. */
-export interface State {}
-
 /** Creates state based on object initial values. */
-export function createState<T extends State = State>(object: T): T;
+export function createState<T extends object = object>(object: T): T;
 
 export type HookCallback = (self: Element) => void | Promise<void>;