+++ src/state/state.ts
@@ -1,9 +1,7 @@
 import { render } from '../dom/render';
 import { _current, _routerGlobalOptions } from '../router/router';
 
-export interface State {}
-
-export function createState<T extends State = State>(obj: T): T {
+export function createState<T extends object = object>(obj: T): T {
   return new Proxy<T>(obj, {
     set(target, property, value, receiver) {
       /**
@@ -13,7 +11,7 @@ export function createState<T extends State = State>(obj: T): T {
        * new object. If value is primitive, there is no
        * need to rerender page as page view will be the same.
        */
-      if (Object.is(target[property as keyof State], value)) {
+      if (Object.is(target[property as keyof T], value)) {
         /**
          * If value is the same as in target there is not need to set
          * value to target, but we must return true to